src/CustomHooks/useResourceManagement.ts
const useResourceManagement = () => {
  const addResource = async (newResource: any) => {
    try {
      return { success: true, message: 'Resource added successfully.' }
    } catch (error) {
      return { success: false, message: 'Failed to add resource.' }
    }
  }

  const editResource = async (resourceId: string, updatedResource: any) => {
    try {
      return { success: true, message: 'Resource updated successfully.' }
    } catch (error) {
      return { success: false, message: 'Failed to update resource.' }
    }
  }

  const deleteResource = async (resourceId: string) => {
    try {
      return { success: true, message: 'Resource deleted successfully.' }
    } catch (error) {
      return { success: false, message: 'Failed to delete resource.' }
    }
  }

  return { addResource, editResource, deleteResource }
}

export default useResourceManagement
